pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "flask-login>=0.6.3",
    "flask>=3.1.0",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.10",
    "flask-migrate>=4.1.0",
    "apscheduler>=3.11.0",
    "python-dateutil>=2.9.0.post0",
    "pytest>=8.3.5",
    "flask-wtf>=1.2.2",
    "requests>=2.32.3",
    "pyzk>=0.9",
    "twilio>=9.5.2",
    "trafilatura>=2.0.0",
    "sqlalchemy>=2.0.40",
    "werkzeug>=3.1.3",
    "beautifulsoup4>=4.13.4",
    "xlsxwriter>=3.2.3",
    "openai>=1.77.0",
]
